Add update ability
With the addition of the Store rule, rules will be able to be updated as
well as applied. This means we need to distinguish between these two
actions. So we add the update method to the Rule class and a boolean to
the State class, and make the necessary tweaks in the tests as well.